pyproject.toml
[project]
name = "memory-vlm"
version = "0.1.0"
description = "Vision-Language Model with in-weight memory augmentation inspired by CLIP"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "timm>=1.0.0",
    "einops>=0.8.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "safetensors>=0.4.0",
    "tensorboard>=2.18.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.18.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.5.0",
    "transformers>=4.45.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "data", "training", "utils"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
